+++ src/commands/index.ts
@@ -160,6 +160,8 @@ import { cacheUpdate } from './moderation/cacheUpdate';
 import { navigraph } from './general/navigraph';
 import { botIssue } from './support/botIssue';
 import { slowMode } from './moderation/slowmode';
+import { econnreset } from './support/econnreset';
+import { pr } from './utils/pr';
 
 const commands: BaseCommandDefinition[] = [
     typeCommand,
@@ -322,6 +324,8 @@ const commands: BaseCommandDefinition[] = [
     navigraph,
     botIssue,
     slowMode,
+    econnreset,
+    pr,
 ];
 
 const commandsObject: { [k: string]: BaseCommandDefinition } = {};