+++ src/commands/support/trythis.ts
@@ -1,18 +1,22 @@
-import { CommandDefinition } from '../../lib/command';
 import { CommandCategory } from '../../constants';
-import { makeEmbed } from '../../lib/embed';
+import { MessageCommandDefinition } from '../../lib/command';
+import { makeEmbed, makeLines } from '../../lib/embed';
 
-export const trythis: CommandDefinition = {
+const trythisDescription = makeLines([
+    'Please try and remove all other mods/liveries from the community folder and test our addon again. This will help rule out mod conflicts.',
+    '',
+    'See [this step](https://docs.flybywiresim.com/fbw-a32nx/support/reported-issues/#test-with-only-the-a32nx-add-on-in-community) in our troubleshooting guide for more details on how to do this easily.',
+]);
+
+const trythisEmbed = makeEmbed({
+    title: 'Try This',
+    description: trythisDescription,
+    footer: { text: 'Report back the result of this test.' },
+});
+
+export const trythis: MessageCommandDefinition = {
     name: 'trythis',
     description: 'Provide basic troubleshooting steps',
     category: CommandCategory.SUPPORT,
-    executor: (msg) => {
-        const trythisEmbed = makeEmbed({
-            title: 'Try This',
-            description: 'Please try and remove all other mods/liveries from the community folder and test our addon again. This will help rule out mod conflicts.',
-            footer: { text: 'Report back the result of this test.' },
-        });
-
-        return msg.channel.send({ embeds: [trythisEmbed] });
-    },
+    genericEmbed: trythisEmbed,
 };