+++ .github/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 Update <small>_ February 2023</small>
 
-- feat: feat: unified command permissions framework (05/02/2023)
+- feat: feat: unified command permissions framework (07/02/2023)
+- feat: ECONNRESET support command (07/02/2023)
+- fix: add link to docs to .trythis command (07/02/2023)
+- feat: PR command (07/02/2023)
 
 Update <small>_ January 2023</small>